[Day 2 ex 02] - adjust the type

+++ Day_2/ex02_bmi_calculator.py
@@ -9,7 +9,7 @@ weight = input("enter your weight in kg: ")
 
 #Write your code below this line 👇
 height = float(height)
-weight = int(weight)
+weight = float(weight)
 
 bmi = (weight) / (height ** 2)
